requirements.txt
streamlit>=1.35.0
openai>=1.52.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.1
duckduckgo-search>=6.2.0
PyMuPDF>=1.24.0
numpy>=1.26.0
